+++ src/main/kotlin/bass/controller/admin/usecase/FindMembersWithRecentCartActivityUseCase.kt
@@ -1,6 +1,6 @@
 package bass.controller.admin.usecase
 
-import bass.dto.ActiveMemberDTO
+import bass.dto.member.ActiveMemberDTO
 
 interface FindMembersWithRecentCartActivityUseCase {
     fun findMembers(): List<ActiveMemberDTO>
