+++ src/main/kotlin/bass/dto/member/ActiveMemberDTO.kt
@@ -1,4 +1,4 @@
-package bass.dto
+package bass.dto.member
 
 import bass.util.ValidationMessages
 import jakarta.validation.constraints.Email
